Revert "Complete Refactoring: Modular Architecture, Async Support, and Code Organization"
