--- __init__.py
@@ -1,7 +1,0 @@
-"""
-Searcharr
-Sonarr, Radarr & Readarr Telegram Bot
-By Todd Roberts
-https://github.com/toddrob99/searcharr
-"""
-__version__ = "3.4.0"
