--- bot/__init__.py
@@ -1,7 +1,0 @@
-"""
-Searcharr
-Sonarr, Radarr & Readarr Telegram Bot
-Bot Module
-By Todd Roberts
-https://github.com/toddrob99/searcharr
-"""
