--- bot/callbacks/base/__init__.py
@@ -1,15 +1,0 @@
-"""
-Searcharr
-Sonarr, Radarr & Readarr Telegram Bot
-Base Callback Utilities
-By Todd Roberts
-https://github.com/toddrob99/searcharr
-"""
-
-# Import all utilities to make them available through the package
-from bot.callbacks.base.navigation import handle_navigation
-from bot.callbacks.base.cancel import handle_cancel
-from bot.callbacks.base.path_selection import check_path_selection
-from bot.callbacks.base.quality_selection import check_quality_selection
-from bot.callbacks.base.tag_selection import handle_tag_selection, process_tags
-from bot.callbacks.base.media_message import update_media_message
