+++ requirements.txt
@@ -1,5 +1,5 @@
 argparse
 requests
-python-telegram-bot
+python-telegram-bot==13.15
 pyyaml
 arrow
