+++ settings-sample.py
